Tracktor from boxes

+++ tracktor/videoflow_contrib/tracktor/tracktor.py
@@ -127,8 +127,8 @@ class TracktorFromBoxes(OneTaskProcessorNode):
             - bboxes: (torch.Tensor) (nb_boxes, 4)
             - scores: (torch.Tensor) (nb_boxes, 1)
         '''
-        bboxes = torch.from_numpy(bounding_boxes[:, [0, 1, 2, 3]].copy())
-        scores = torch.from_numpy(bounding_boxes[:, [4]].copy())
+        bboxes = torch.from_numpy(bounding_boxes[:, [0, 1, 2, 3]].astype('float').copy())
+        scores = torch.from_numpy(bounding_boxes[:, [4]].astype('float').copy())
         return bboxes, scores
     
     def open(self):
